refactor(structure): reorganize code into modules
