+++ jest.config.ts
@@ -9,6 +9,9 @@ const config: Config.InitialOptions = {
   moduleNameMapper: {
     '^@src/(.*)$': '<rootDir>/src/$1',
     '^@test/(.*)$': '<rootDir>/test/$1',
+    '^@core/(.*)$': '<rootDir>/src/modules/core/$1',
+    '^@strict-store/(.*)$': '<rootDir>/src/modules/strict-store/$1',
+    '^@strict-json/(.*)$': '<rootDir>/src/modules/strict-json/$1',
   },
   testPathIgnorePatterns: ['/node_modules/'],
 };
