--- src/app/ports/serializer.port.ts
@@ -1,7 +1,0 @@
-import { Persistable } from '@src/domain/entities/persistable.entity';
-import { StoreKey } from '@src/domain/entities/store-key.entity';
-
-export interface SerializerPort {
-  parse<T extends Persistable>(s: string): T;
-  stringify<T extends StoreKey<Persistable>>(value: T['__type']): string;
-}
