--- src/app/ports/storage-provider.port.ts
@@ -1,6 +1,0 @@
-import { StoreType } from '@src/domain/entities/store-type.entity';
-import { KeyValueStoragePort } from '@src/app/ports/key-value-storage.port';
-
-export interface StorageProviderPort {
-  get(type: StoreType): KeyValueStoragePort;
-}
