--- src/domain/constants/key-pattern.constant.ts
@@ -1,3 +1,0 @@
-import { KEY_PREFIX } from '@src/domain/constants/key-prefix.contant';
-
-export const KEY_PATTERN = new RegExp(`^${KEY_PREFIX}/([^:]+):(.+)$`);
