--- src/domain/policies/typed-array.policy.ts
@@ -1,6 +1,0 @@
-import { TypedArray } from '@src/domain/entities/typed-array.entity';
-
-export const typedArrayPolicy = {
-  isTypedArray: (val: unknown): val is TypedArray =>
-    ArrayBuffer.isView(val) && !(val instanceof DataView),
-};
