+++ src/modules/core/constants/key-pattern.constant.ts
@@ -1,0 +1,3 @@
+import { KEY_PREFIX } from '@core/constants/key-prefix.contant';
+
+export const KEY_PATTERN = new RegExp(`^${KEY_PREFIX}/([^:]+):(.+)$`);
