+++ src/modules/core/entities/complex-type.entity.ts
@@ -1,0 +1,9 @@
+import { Persistable } from '@core/entities/persistable.entity';
+
+export type ComplexTypeNames = 'bigint' | 'set' | 'map' | 'typedArray';
+
+export type ComplexTypeData = {
+  __type: ComplexTypeNames;
+  value: Persistable;
+  subtype?: string;
+};
