+++ src/modules/core/entities/deep-partial.entity.ts
