+++ src/modules/core/entities/store-type.entity.ts
