+++ src/modules/core/entities/typed-array.entity.ts
