+++ src/modules/core/policies/typed-array.policy.ts
@@ -1,0 +1,6 @@
+import { TypedArray } from '@core/entities/typed-array.entity';
+
+export const typedArrayPolicy = {
+  isTypedArray: (val: unknown): val is TypedArray =>
+    ArrayBuffer.isView(val) && !(val instanceof DataView),
+};
