+++ src/modules/strict-json/app/ports/complex-type.port.ts
@@ -1,0 +1,14 @@
+import type { Persistable, NativePersistable } from '@core/entities/persistable.entity';
+import type { ComplexTypeData } from '@core/entities/complex-type.entity';
+
+export interface ComplexTypeCodecPort {
+  /**
+   * Returns ComplexTypeData if value is a supported complex type, otherwise null.
+   */
+  encode(value: Persistable): ComplexTypeData | null;
+
+  /**
+   * Returns a complex runtime value if the payload is a ComplexTypeData, otherwise null.
+   */
+  decode(value: NativePersistable | ComplexTypeData): Persistable | null;
+}
