+++ src/modules/strict-store/app/ports/key-value-storage.port.ts
